==> Machine_Learning/Driver_demand_prediction/code/1. Data Analysis and Preprocessing.ipynb <==
import pandas as pd
import numpy as np
import pandas_profiling
from datetime import timedelta

!ls ./../data/

## Driver Data Analysis

df_drivers = pd.read_csv('./../data/drivers.csv')

print("Intially Number of Rows in Driver.csv", len(df_drivers))

df_drivers.head(10)

## Driver Profile Data contains Duplicates
df_drivers[df_drivers.duplicated(subset=['driver_id'],keep=False)]

### Same driver_id cannot have different profile data
#### Considering first occurence to be correct. 
#### We have 2497 unique driver_id: profiles

## Keeping first occurence
df_drivers.drop_duplicates(subset=['driver_id'], inplace = True)

df_drivers.head(10)

# Info of Driver Dataset
df_drivers.info()

#Count missing values
np.count_nonzero(df_drivers.isnull().values)

df_drivers.reset_index(inplace = True, drop = True)

profile = df_drivers.profile_report(title='Driver DataSet Analysis')
profile.to_file(output_file="../data_analysis_EDA/driver_data.html")
df_drivers.to_csv('../processed_data_and_models/driver.csv',index = False)

## Pings Data Analysis and Preprocessing

df_pings = pd.read_csv('./../data/pings.csv')

print("Intially Number of Rows in pings.csv", len(df_pings))

df_pings.head(10)

df_pings.info()

df_pings.sort_values(by = ['driver_id', 'ping_timestamp']).reset_index(drop = True)

df_pings.head(10)

# Count missing values
np.count_nonzero(df_pings.isnull().values)

# Check Data contains duplicates 
len(df_pings[df_pings.duplicated(keep=False)])

df_pings.drop_duplicates(subset= ['driver_id', 'ping_timestamp'], inplace = True)

df_pings['ping_date']= pd.to_datetime(df_pings['ping_timestamp'], unit= 's')

## Converting UTC Time to IST by adding 05:30 time difference
df_pings['ping_date'] = df_pings['ping_date']+pd.to_timedelta('05:30:00.00000')

df_pings['ping_date'] = df_pings['ping_date'].dt.date

## We know that: every 15 seconds indicating that driver is available.
## Methodology of Lag and Lead will be used here. 
### difference(t) = observed(t) - observed(t-1)
### If difference between two ping_timestamps is less than `ping_threshold` means driver was available for that time. 

### To get lag of interval =1, on grouped date by (driverid and ping_date) we can used shift() or diff()

# grouped difference on driverid and ping_date
df_pings['diff'] = df_pings.groupby(['driver_id', 'ping_date'])['ping_timestamp'].diff().fillna(0)

df_pings[(df_pings['driver_id']==111556)]

df_pings['diff'].value_counts().head(10)

### To get hours of driver being available for a day we sum all difference timestamps which are within PING_THRESHOLD 

PING_THRESHOLD = 60

df_pings[df_pings['diff']<=PING_THRESHOLD].head()

data = pd.DataFrame(df_pings[df_pings['diff']<=PING_THRESHOLD].groupby(by=['driver_id','ping_date'])['diff'].agg('sum')).reset_index()
data.columns = ['driver_id', 'date','available_hours']
# Secs to Hours
data['available_hours'] = round(data['available_hours']*(1.0/3600.0), 1)

data.head(8)

## Fill in missing Date For Driver with Available_hours = 0
data.date = pd.to_datetime(data.date)
data = data.set_index(
    ['date', 'driver_id']
).unstack(
    fill_value=0
).asfreq(
    'D', fill_value=0
).stack().sort_index(level=1).reset_index()

data

### DayofWeeks: where 0: Monday and 6: Sunday

data['dayofweek'] = data['date'].dt.dayofweek
data.head()

### For some drivers there is ping data for even 22 June, removing that. 
#### Given in Problem Statement Data is provided from 1st June to 21st June

data= data[data.date!='2017-06-22']

data.reset_index(drop = True, inplace = True)

profile = data.profile_report(title='Driver_Pings Data Analysis')

profile.to_file(output_file="../data_analysis_EDA/driver_pings_hours.html")

data.to_csv('../processed_data_and_models/Pings_Grouped_Data.csv',index = False)

pd.read_csv('./../data/test.csv').profile_report(title = 'Drivers_Test Data').to_file(output_file="../data_analysis_EDA/test_driver.html")
