==> Machine_Learning/Driver_demand_prediction/code/2. Preparing_TrainingData.ipynb <==
## Dataset Insights

### DriversProfile.csv
+ No NULL or Missing Values
+ Duplicate Values subset=[driver_id], count: 3
+ 2497: Unique Driver Ids
+ Driver Age varies from 18 to 75 | Median at 31years
+ Gender column is categorical with two classes Male and Female; Male= 1894, Female= 606
+ Age and Number of Children has +ve correlation

### Pings_Data
+ 2 Columns | Driverid and pingtimestamp
+ Data Contains Duplicate Values: 79086 count
+ No NaN values found in Data.
+ 2480: Unique Driver Ids
+ No considering: Removing 22nd June Data

### Processed Pings Data to get Number of hours a Driver is Available
+ Data rows: Unique Driver Ids * 21days = 2480 * 21 = 52080

### Final Training Dataset 
+ Should Have: 2497 driver_id and 2497*21=> 52437 data rows

import pandas as pd
import numpy as np
import pandas_profiling

!ls

df_pings = pd.read_csv('./../data/pings.csv')

## Count of Unique drivers in Pings.csv
len(np.unique(df_pings.driver_id.values))

## Processed Driver Hours Data
driver_hours = pd.read_csv('../processed_data_and_models/Pings_Grouped_Data.csv')

## VERIFIED | Data rows: Unique Driver Ids * 21days = 2480 * 21 = 52080
assert 52080 == len(driver_hours)

driver_profile = pd.read_csv('../processed_data_and_models/driver.csv')

driver_profile.head(10)

## Driver_Id for which we don't have PingData or they were not available 
set(np.unique(driver_profile.driver_id.values)) - set(np.unique(driver_hours.driver_id.values))

### Merge Data | Driver Profile and Driver Hours on = [driver_id]

## Left Join
dataset = pd.merge(driver_profile, driver_hours, on = ['driver_id'], how = 'left')

dataset[dataset.driver_id == 230923]

dataset['date'].fillna('2017-06-01',inplace = True)

## Fill in missing Date For Driver with Available_hours = 0
dataset.date = pd.to_datetime(dataset.date)
dataset = dataset.set_index(
    ['date', 'driver_id']
).unstack().fillna(method = 'ffill').asfreq(
    'D'
).stack().sort_index(level=1).reset_index()

dataset['dayofweek'].fillna(dataset['date'].dt.dayofweek, inplace = True)
dataset['available_hours'].fillna(0,inplace = True)

## Verifying Data of Driver which was missing
dataset[dataset.driver_id == 998740]

#### Dataset len = Number of Drivers * 21 days => 2497*21 => 52437

assert len(dataset) == 52437

### Label Encoding: Replace Male: 0 and Female: 1

dataset['gender'].replace({'FEMALE':1, 'MALE':0}, inplace = True)

### Encoding-> 1: Weekend  0: Weekday

dataset['weekend'] = dataset['dayofweek'].apply(lambda x: 0 if x < 5 else 1)

### Training Data is Ready

#### Data Analysis of Training Data

dataset.to_csv('../processed_data_and_models/Training_Dataset.csv',index = False)

profile = dataset.profile_report(title='Driver Training Data')

profile.to_file('../data_analysis_EDA/Driver_Training.html')

### Preparing Test Data

test_data = pd.read_csv('./../data/test.csv')

test_data['date'] = pd.to_datetime(test_data['date']) 
test_data['dayofweek'] = test_data['date'].dt.dayofweek
test_data['weekend'] = test_data['dayofweek'].apply(lambda x: 0 if x < 5 else 1)

test = pd.merge(test_data,driver_profile, on = ['driver_id'])

test['gender'].replace({'FEMALE':1, 'MALE':0}, inplace = True)
test.to_csv('../processed_data_and_models/driver_test.csv',index = False)
